split cli.py by subcommands
Reduces file size below 1000 lines.
