+++ nominatim/clicmd/__init__.py
@@ -1,0 +1,8 @@
+"""
+Subcommand definitions for the command-line tool.
+"""
+
+from .replication import UpdateReplication
+from .api import APISearch, APIReverse, APILookup, APIDetails, APIStatus
+from .index import UpdateIndex
+from .refresh import UpdateRefresh
